+++ src/CausalityTools.jl
@@ -56,12 +56,11 @@ module CausalityTools
     # Update messages:
     using Scratch
     display_update = true
-    version_number = "2.3.1"
+    version_number = "2.6.0"
     update_name = "update_v$(version_number)"
     update_message = """
     \nUpdate message: CausalityTools v$(version_number)\n
-    - The `MesnerShalisi` estimator is now deprecated and renamed to `MesnerShalizi` (with
-    correct spelling).
+    - New causal graph inference algorithm: `PC`.
     """
 
     if display_update