+++ changelog.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## 2.6.0
+
+- New causal graph inference algorithm: `PC`.
+
+## 2.5.0
+
+- New independence test: `CorrTest`, based on (partial) correlations.
+
+## 2.4.0
+
+- Added partial distance correlation measure. To compute it, simply provide a
+    third input argument to `distance_correlation`.
+- `DistanceCorrelation` is now compatible with both `SurrogateTest` and
+    `LocalPermutationTest` in its conditional form.
+
 ## 2.3.1
 
 - The `MesnerShalisi` estimator is now deprecated and renamed to `MesnerShalizi` (with